src: Move api_headers to common/
