+++ src/assets/MeshAsset.h
@@ -18,7 +18,7 @@
 #include "assets/Asset.h"
 #include "gpu/GpuBuffer.h"
 #include "gpu/GpuInstance.h"
-#include "src/api_headers.h"
+#include "common/api_headers.h"
 
 namespace mondradiko {
