+++ src/assets/TextureAsset.h
@@ -17,7 +17,7 @@
 #include "assets/Asset.h"
 #include "gpu/GpuImage.h"
 #include "gpu/GpuInstance.h"
-#include "src/api_headers.h"
+#include "common/api_headers.h"
 
 namespace mondradiko {
