+++ src/filesystem/AssimpIOStream.h
@@ -15,7 +15,7 @@
 #include <vector>
 
 #include "filesystem/Filesystem.h"
-#include "src/api_headers.h"
+#include "common/api_headers.h"
 
 namespace mondradiko {
