+++ src/gpu/GpuImage.h
@@ -14,7 +14,7 @@
 
 #include "gpu/GpuBuffer.h"
 #include "gpu/GpuInstance.h"
-#include "src/api_headers.h"
+#include "common/api_headers.h"
 
 namespace mondradiko {
