+++ src/log/log.h
@@ -9,7 +9,7 @@
  *
  */
 
-#include "./api_headers.h"
+#include "common/api_headers.h"
 
 #ifndef SRC_LOG_LOG_H_
 #define SRC_LOG_LOG_H_
