+++ src/renderer/MeshPipeline.h
@@ -27,7 +27,7 @@
 #include "gpu/GpuDescriptorSetLayout.h"
 #include "gpu/GpuDescriptorPool.h"
 #include "gpu/GpuInstance.h"
-#include "src/api_headers.h"
+#include "common/api_headers.h"
 
 namespace mondradiko {
